+++ src/util/toolbox.hpp
@@ -1,0 +1,26 @@
+#ifndef __TOOLBOX_HPP
+#define __TOOLBOX_HPP
+
+#include <chrono>
+#include <iomanip>
+#include <sstream>
+#include <string>
+#include <unordered_map>
+
+#include "string_tools.hpp"
+#include "../conf.hpp"
+
+bool doesFileExist(const std::string&, const bool);
+bool doesDirectoryExist(const std::string&, const bool);
+int loadErrorDoc(const int, const std::string&, std::string&);
+int loadConfHeaders(std::unordered_map<std::string, std::string>&);
+
+void formatFileSize(size_t, std::string&);
+void formatDate(const std::chrono::system_clock::duration, std::string&);
+
+int loadDirectoryListing(std::string&, const std::string&, const std::string&);
+
+// Debug profiling
+long long debug_getTimestamp();
+
+#endif